requirements.txt
pythonnet>=3.0.0
pyserial>=3.0.0
pytest
pytest-mock